+++ server/controllers/apiTradeLoan.js
@@ -31,20 +31,20 @@ exports.readAllTradeLoan = async (req, res) => {
   exports.createTradeLoan = async (req, res) => {
     const tradeLoan = new TradeLoan();
 
-    tradeLoan.userOfering = req.body.userOfering;
+    tradeLoan.userOfering = req.user.id;
     tradeLoan.userTaking = req.body.userTaking;
     tradeLoan.product = req.body.product;
    tradeLoan.publishingDate = req.body.publishingDate;
    tradeLoan.returnDate = req.body.returnDate;
      
     try {
-        const userOfering = await User.findById({_id:req.body.userOfering});
+        const userOfering = await User.findById({_id:req.user.id});
         if (userOfering == null) res.status(404).json({error:"userOfering not found"});
       
         const userTaking = await User.findById({_id:req.body.userTaking});
         if (userTaking == null) res.status(404).json({error:"userTaking not found"});
 
-        if (req.body.userOfering == req.body.userTaking) res.status(404).json({error:"userTaking == userOfering"});
+        if (req.user.id == req.body.userTaking) res.status(404).json({error:"userTaking == userOfering"});
        
         const product = await Product.findById({_id:req.body.product});
         if (product == null) res.status(404).json({error:"product not found"});
@@ -71,13 +71,18 @@ exports.readAllTradeLoan = async (req, res) => {
       console.log('Can not create the tradeLoan');
     }
 };
-  
-  exports.updateTradeLoan = async (req, res) => {
-  
-};
 
   exports.deleteTradeLoan = async (req, res) => {
   
+    /*if ('ADMIN' != req.user.role) {
+      res.status(401).json({error: "Do not have permission"})
+      return;
+    }*/
+    const user = await User.findbyId(req.user.id);
+    if (user.role != 'ADMIN'){
+      res.status(401).json({error: "Do not have permission"})
+      return;
+    }
       try {
         const tradeLoan = await TradeLoan.findByIdAndDelete({_id: req.params.id});
